Created suitable_lobby.py
Moved everything relevant to finding a suitable lobby into a new file.
Also refactored find_suitable method
